+++ iMira/CodeFenceInfo.swift
@@ -1,6 +1,6 @@
 //
 //  CodeFenceInfo.swift
-//  md-preview
+//  iMira
 //
 
 import Foundation