+++ iMira/MarkdownFrontmatter.swift
@@ -1,6 +1,6 @@
 //
 //  MarkdownFrontmatter.swift
-//  md-preview
+//  iMira
 //
 
 import Foundation