+++ tests/swift-tests/Sources/MarkdownHelpers/EscapingHTMLFormatter.swift
@@ -1,1 +1,1 @@
-../../../../md-preview/EscapingHTMLFormatter.swift
+../../../../iMira/EscapingHTMLFormatter.swift